rename ReportScreenContext and extract ActionListContext
