+++ src/hooks/useReportScrollManager/index.native.ts
@@ -1,6 +1,6 @@
 // eslint-disable-next-line no-restricted-imports
 import type {ScrollView} from 'react-native';
-import {useActionListContext} from '@pages/inbox/ReportScreenContext';
+import {useActionListContext} from '@pages/inbox/ActionListContext';
 import type ReportScrollManagerData from './types';
 
 function useReportScrollManager(): ReportScrollManagerData {
