+++ src/pages/inbox/ReactionListWrapper.tsx
@@ -1,9 +1,9 @@
 import React, {useEffect, useRef, useState} from 'react';
 import type {SyntheticEvent} from 'react';
 import {Dimensions} from 'react-native';
+import {ReactionListContext} from './ReactionListContext';
+import type {ReactionListAnchor, ReactionListContextType, ReactionListEvent} from './ReactionListContext';
 import PopoverReactionList from './report/ReactionList/PopoverReactionList';
-import {ReactionListContext} from './ReportScreenContext';
-import type {ReactionListAnchor, ReactionListContextType, ReactionListEvent} from './ReportScreenContext';
 
 type AnchorPosition = {horizontal: number; vertical: number};
