+++ src/pages/inbox/report/ReactionList/PopoverReactionList.tsx
@@ -4,7 +4,7 @@ import PopoverWithMeasuredContent from '@components/PopoverWithMeasuredContent';
 import useCurrentUserPersonalDetails from '@hooks/useCurrentUserPersonalDetails';
 import useOnyx from '@hooks/useOnyx';
 import {getEmojiReactionDetails, mergeReactionsByEmoji} from '@libs/EmojiUtils';
-import type {ReactionListAnchor} from '@pages/inbox/ReportScreenContext';
+import type {ReactionListAnchor} from '@pages/inbox/ReactionListContext';
 import ONYXKEYS from '@src/ONYXKEYS';
 import {multiPersonalDetailsSelector} from '@src/selectors/PersonalDetails';
 import type {PersonalDetails} from '@src/types/onyx';
