+++ src/pages/inbox/report/ReportActionCompose/useComposerSubmit.ts
@@ -14,8 +14,8 @@ import {rand64} from '@libs/NumberUtils';
 import {addDomainToShortMention} from '@libs/ParsingUtils';
 import {startSpan} from '@libs/telemetry/activeSpans';
 import {generateAccountID} from '@libs/UserUtils';
+import {useActionListContext} from '@pages/inbox/ActionListContext';
 import {useAgentZeroStatusActions} from '@pages/inbox/AgentZeroStatusContext';
-import {useActionListContext} from '@pages/inbox/ReportScreenContext';
 import {setIsComposerFullSize} from '@userActions/Report';
 import CONST from '@src/CONST';
 import ONYXKEYS from '@src/ONYXKEYS';
