+++ tests/perf-test/ReportActionsList.perf-test.tsx
@@ -6,8 +6,9 @@ import OnyxListItemProvider from '@components/OnyxListItemProvider';
 import type Navigation from '@libs/Navigation/Navigation';
 import navigationRef from '@libs/Navigation/navigationRef';
 import {setHasRadio} from '@libs/NetworkState';
+import {ActionListContext} from '@pages/inbox/ActionListContext';
+import {ReactionListContext} from '@pages/inbox/ReactionListContext';
 import ReportActionsList from '@pages/inbox/report/ReportActionsList';
-import {ActionListContext, ReactionListContext} from '@pages/inbox/ReportScreenContext';
 import {AttachmentModalContextProvider} from '@pages/media/AttachmentModalScreen/AttachmentModalContext';
 import ComposeProviders from '@src/components/ComposeProviders';
 import {LocaleContextProvider} from '@src/components/LocaleContextProvider';
